Save audio/subtitle preferences immediately
Immediately saves the audio and subtitle preference when it is changed,
instead of showing the save request through the modal dialog box.

This solves different problems:
-User misunderstanding caused by double saving of preferences (actual
change via kodi and request to save from the addon)
-If you don't press Ok within 5 seconds from the appearance of the modal
window, the saving doesn't happen and you have to change the preferences
twice more in order to save the setting correctly.
-Problems of interaction with Kodi's UI when changing the subtitle or
audio track from the Kodi player menu