Update workaround subtitle position problem
The current version of inputstream.adaptive has not yet been fixed to better manage
the webvtt profile.

symptom: when there are more lines of subtitles displayed at the same
time the position of the subtitles is wrong and reversed.

cause: screen resolutions below 1080P or display Kodi in window mode causes
the mentioned problem