msl_handler: some cleaning for hdcp
With manifest v2, the request for the hdcp version is made
through the manifest. Therefore in the response there are no more hdpc
tags for each video stream, so the hdcp tag can currently be removed from the
mpd file.
To enable 4k it is still mandatory to enable Override HDCP on
inputstream.adaptive in order to force the highest resolution.