Fixed loading of menus in the children's profile
The main cause of the missing menu loading was due to to the "Characters" menu which is currently not supported and should be implemented.

In the Lolomo answer, in the lists elements, the Id of this menu is positioned with index 0. As index 0, it was read first, causing an error by blocking the loading of the other menus for an unmanaged part of the code.

For future implementation, note the validation in VideoId.py, which marks the id of this menu as UNSPECIFIED.