Filterout unsuitable contexts
In some rare cases, when you try to access a profile you will see the error:

'Can only construct a VideoId from a show or movie item'
InvalidVideoId: Can only construct a VideoId from a show or movie item

because netflix has introduced a new context (showAsARow) in lolomo that contains only 'summary' elements of the type 'episode' instead of the types 'movie' or 'show', for our addon structure this context is not useful to us and only causes problems to manage it.

Apparently this new context (whose content is changed from time to time by netflix) is added to the lolomo response only if a certain TV series is not contained in the "My list".