Fix esn key handshake on session initialization
In the many prevoius changes i made a mistake,
the esn after initialization should always be maintained
because the key handshake was created with that esn.
To keep in mind that however that the esn from the web login and
esn used locally are different.
Also there was a check for the change of esn in the login was incorrect.