Modified login persistence system
previously to check if the login was still valid, a request was made to
get the profile data at every interaction with the interface, causing
slowdowns in loading and multiple unnecessary requests.
now only the expiration dates of the cookies are verified and in case of
error uses the previous system.