Small fixes to Dutch translations
Not a lot of changes.

One problem is that the English translation was changed, but the
translation files were not updated for this. This may be problematic for
others as well.

I also translated My list to 'Mijn lijst' using quotes so it looks
much better. In English it is not done consistently, and not
consistently with single quotes.