ESN changes see details
This change mainly implements an option to force identify a L1 device as L3
It is needed to try to allow video playback on some devices that have L1
specs, but have a broken/non regular Widevine library.

It also include some improvements to  ESN documetation,
and move to a separated module the ESN generation function.