Included option to enable nf watched status sync to library (one way)
This feature is added one-way, i.e. it only sends changes to the netflix service but does not update the Kodi library if the watched statuses are changed externally.

The NF metadata API currently seem does not provide real-time updated data,
so you cannot use it to synchronize the watched status of the videos.

And there is also a second problem,
the profile switch is only executed at the beginning of the sync, and immediately restored (nf manages one profile at a time).
So the metadata requests are obtained associated with the last selected profile, and this will cause a mix of bookmark data from other profiles.