[Readme] Add warning for suspended development
After years of maintenance and development of this add-on i am sorry to have to declare its development suspended indefinitely.
Unfortunately to date i am unable to address the problem regarding video playback with my skills alone. If someone is able in the future to fix it, time permitting i can provide support to make new releases. Thanks for understanding