fix: update auth_url validation method
It seems netflix now returns tokens which have a length of 77
characters.
The new token is now composed of three parts separated by dots:
- a prefix: c1
- a timestamp in milliseconds
- a base64 encoded string

This code updates the validation mecanism to take this into account.
fix #1740

This code also fixes the regex which extracts the client version
number.
The default fallback version of the client has been updated as well.